+++ src/index.html
@@ -5,7 +5,8 @@
   <meta name="viewport" content="width=device-width, initial-scale=1.0">
   <meta http-equiv="X-UA-Compatible" content="ie=edge">
   <title>Social Veg</title>
-  <link rel="stylesheet" href="./lib/view/logInAndSignUp/singUp.css" />  
+  <link rel="stylesheet" href="./lib/view/logInAndSignUp/singUp.css" /> 
+  <link rel="stylesheet" href="./lib/view/wall/wall.css" /> 
   <link rel="preconnect" href="https://fonts.gstatic.com">
   <link href="https://fonts.googleapis.com/css2?family=Shrikhand&display=swap" rel="stylesheet">
   <link rel="preconnect" href="https://fonts.gstatic.com">
@@ -13,11 +14,10 @@
 </head>
 <body>
   <!-- Donde se imprimiera todo los templace -->
-
-  <header id=menu>
-  </header>
   
   <div id="root"> 
+  </div>
+  <div id="modal"> 
   </div>
 
   </body>
